🎨 Rename : "기본 레시피 디렉토리명 및 경로 수정"
- 기존 categorysearch를 defaultcocktail로 변경하였습니다.
 - 디렉토리, 경로, 템플릿이 수정되었습니다.

Ref : #16
