+++ src/main/java/alcoholboot/toastit/feature/defaultcocktail/controller/CocktailController.java
@@ -1,12 +1,11 @@
-package alcoholboot.toastit.feature.categorysearch.controller;
+package alcoholboot.toastit.feature.defaultcocktail.controller;
 
-import alcoholboot.toastit.feature.categorysearch.domain.Cocktail;
-import alcoholboot.toastit.feature.categorysearch.service.CocktailService;
+import alcoholboot.toastit.feature.defaultcocktail.domain.Cocktail;
+import alcoholboot.toastit.feature.defaultcocktail.service.CocktailService;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -15,20 +14,11 @@ import java.util.List;
 import java.util.Optional;
 
 @Controller
-//@RestController
 @RequiredArgsConstructor
 @RequestMapping("/cocktails")
 public class CocktailController {
     private final CocktailService cocktailService;
 
-    // 레시피 탐색
-    /*@GetMapping("/all")
-    public ResponseEntity<Page<Cocktail>> getAllCocktails(@RequestParam(defaultValue = "0") int page) {
-
-        Page<Cocktail> cocktails = cocktailService.getAllCocktailsPaged(PageRequest.of(page, 20)); // 20개씩 페이징
-
-        return ResponseEntity.ok(cocktails);
-    }*/
     @GetMapping("/all")
     public String getAllCocktails(
             @RequestParam(defaultValue = "0") int page,
@@ -38,15 +28,9 @@ public class CocktailController {
         model.addAttribute("cocktails", cocktails);
         model.addAttribute("currentPage", page);
         model.addAttribute("totalPages", cocktails.getTotalPages());
-        return "/feature/categorysearch/cocktailList";
+        return "/feature/defaultcocktail/cocktailList";
     }
 
-    /*@GetMapping("/all/ingredient")
-    public ResponseEntity<List<Cocktail>> getCocktailsByIngredient(
-            @RequestParam String ingredient) {
-        List<Cocktail> cocktails = cocktailService.getCocktailsByIngredient(ingredient);
-        return ResponseEntity.ok(cocktails);
-    }*/
     @GetMapping("/all/ingredient")
     public String getCocktailsByIngredient(
             @RequestParam(defaultValue = "0") int page,
@@ -57,16 +41,9 @@ public class CocktailController {
         model.addAttribute("cocktails", cocktails);
         model.addAttribute("page", page);
         model.addAttribute("ingredient", ingredient);
-        return "feature/categorysearch/cocktailIngredient";
+        return "feature/defaultcocktail/cocktailIngredient";
     }
 
-
-    //    @GetMapping("/all/glass")
-//    public ResponseEntity<List<Cocktail>> getCocktailsByGlass(
-//            @RequestParam String glass) {
-//        List<Cocktail> cocktails = cocktailService.getCocktailsByGlass(glass);
-//        return ResponseEntity.ok(cocktails);
-//    }
     @GetMapping("/all/glass")
     public String getCocktailsByGlass(
             @RequestParam(defaultValue = "0") int page,
@@ -77,16 +54,9 @@ public class CocktailController {
         model.addAttribute("cocktails", cocktails);
         model.addAttribute("page", page);
         model.addAttribute("glass", glass);
-        return "feature/categorysearch/cocktailGlass";
+        return "feature/defaultcocktail/cocktailGlass";
     }
 
-
-    //    @GetMapping("/all/type")
-//    public ResponseEntity<List<Cocktail>> getCocktailsByType(
-//            @RequestParam String type) {
-//        List<Cocktail> cocktails = cocktailService.getCocktailsByType(type);
-//        return ResponseEntity.ok(cocktails);
-//    }
     @GetMapping("/all/type")
     public String getCocktailsByType(
             @RequestParam(defaultValue = "0") int page,
@@ -97,19 +67,9 @@ public class CocktailController {
         model.addAttribute("cocktails", cocktails);
         model.addAttribute("page", page);
         model.addAttribute("type", type);
-        return "feature/categorysearch/cocktailType";
+        return "feature/defaultcocktail/cocktailType";
     }
 
-    /*@GetMapping("/all/mult")
-    public ResponseEntity<List<Cocktail>> getCocktailsByMult(
-            @RequestParam(required = false) String ingredient,
-            @RequestParam(required = false) String glass,
-            @RequestParam(required = false) String type,
-            Model model){
-        List<Cocktail> cocktails = cocktailService.getCocktailsByFilter(ingredient, glass, type);
-        model.addAttribute("cocktails", cocktails);
-        return ResponseEntity.ok(cocktails);
-    }*/
     @GetMapping("/all/mult")
     public String getCocktailsByMult(
             @RequestParam(defaultValue = "0") int page,
@@ -124,16 +84,9 @@ public class CocktailController {
         model.addAttribute("glass", glass);
         model.addAttribute("type", type);
 
-        return "feature/categorysearch/cocktailComplex";
+        return "feature/defaultcocktail/cocktailComplex";
     }
 
-    //    @GetMapping("/{id}")
-//    public ResponseEntity<Optional<Cocktail>> getCocktailById(
-//            @PathVariable String id) {
-//
-//        Optional<Cocktail> cocktails = cocktailService.getCocktailById(new ObjectId(id));
-//        return ResponseEntity.ok(cocktails);
-//    }
     @GetMapping("/id")
     public String getCocktailById(
             @RequestParam("id") String id,
@@ -141,6 +94,6 @@ public class CocktailController {
         Optional<Cocktail> cocktail = cocktailService.getCocktailById(new ObjectId(id));
         model.addAttribute("cocktail", cocktail);
 
-        return "feature/categorysearch/cocktailDetails";
+        return "feature/defaultcocktail/cocktailDetails";
     }
 }
