+++ src/main/resources/templates/feature/defaultcocktail/cocktailDetails.html
