+++ src/main/resources/templates/feature/defaultcocktail/cocktailGlass.html
