+++ src/main/resources/templates/feature/defaultcocktail/cocktailIngredient.html
