+++ src/main/resources/templates/feature/defaultcocktail/cocktailList.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link th:href="@{/css/fragment/navbar-fragment.css}" rel="stylesheet">
-    <link th:href="@{/css/categorysearch/categorysearch.css}" rel="stylesheet">
+    <link th:href="@{/css/defaultcocktail/defaultcocktail.css}" rel="stylesheet">
     <title>칵테일 목록</title>
 </head>
 <body>
