+++ src/main/resources/templates/feature/defaultcocktail/cocktailType.html
