+++ src/test/java/alcoholboot/toastit/feature/defaultcocktail/service/impl/CocktailServiceImplTest.java
@@ -1,8 +1,8 @@
-package alcoholboot.toastit.feature.categorysearch.service.impl;
+package alcoholboot.toastit.feature.defaultcocktail.service.impl;
 
-import alcoholboot.toastit.feature.categorysearch.domain.Cocktail;
-import alcoholboot.toastit.feature.categorysearch.repository.CocktailRepository;
-import alcoholboot.toastit.feature.categorysearch.service.CocktailService;
+import alcoholboot.toastit.feature.defaultcocktail.domain.Cocktail;
+import alcoholboot.toastit.feature.defaultcocktail.repository.CocktailRepository;
+import alcoholboot.toastit.feature.defaultcocktail.service.CocktailService;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
 import org.junit.jupiter.api.Test;
