+++ src/test/java/alcoholboot/toastit/feature/user/service/LikeServiceImplTest.java
@@ -3,7 +3,7 @@
 //import static org.junit.jupiter.api.Assertions.*;
 //import static org.mockito.Mockito.*;
 //
-//import alcoholboot.toastit.feature.categorysearch.entity.CocktailEntity;
+//import alcoholboot.toastit.feature.defaultcocktail.entity.CocktailEntity;
 //import alcoholboot.toastit.feature.user.domain.Like;
 //import alcoholboot.toastit.feature.user.domain.User;
 //import alcoholboot.toastit.feature.user.entity.LikeEntity;
